+++ src/components/main-screen/main-screen.jsx
@@ -2,7 +2,7 @@ import React, {Fragment} from 'react';
 import {Header} from '../header/header.jsx';
 import MoviesList from '../movies-list/movies-list.jsx';
 import MoviesFilter from '../movies-filter/movies-filter.jsx';
-import {withActiveItem} from '../../hoc/withActiveItem.jsx';
+import {withActiveItem} from '../../hocs/with-active-item/with-active-item.jsx';
 
 const MoviesListWrapped = withActiveItem(MoviesList);
 const MoviesFilterWrapped = withActiveItem(MoviesFilter);
