+++ src/components/movie-card/movie-card.jsx
@@ -1,9 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {VideoPlayer} from '../video-player/video-player.jsx';
+import {withVideo} from '../../hocs/with-video/with-video.jsx';
+
+const VideoPlayerWrapped = withVideo(VideoPlayer);
 
 export const MovieCard = (props) => {
-  const {id, name, poster, src, onMouseEnter, onMouseLeave, onClick, isPlaying = false} = props;
+  const {
+    id,
+    name,
+    poster,
+    src,
+    onMouseEnter,
+    onMouseLeave,
+    onClick,
+    isPlaying = false
+  } = props;
   const _onMouseEnter = () => onMouseEnter(id);
   const _onClick = (evt) => {
     evt.preventDefault();
@@ -11,9 +23,12 @@ export const MovieCard = (props) => {
   };
 
   return (
-    <article className="small-movie-card catalog__movies-card" onMouseEnter={_onMouseEnter} onMouseLeave={onMouseLeave}>
+    <article
+      className="small-movie-card catalog__movies-card"
+      onMouseEnter={_onMouseEnter}
+      onMouseLeave={onMouseLeave}>
       <div className="small-movie-card__image">
-        <VideoPlayer poster={poster} src={src} isPlaying={isPlaying} muted={true}/>
+        <VideoPlayerWrapped poster={poster} src={src} isPlaying={isPlaying} muted={true}/>
       </div>
       <h3 className="small-movie-card__title" onClick={_onClick}>
         <a className="small-movie-card__link" >{name}</a>
