+++ src/components/movies-list/movies-list.jsx
@@ -1,7 +1,7 @@
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {MovieCard} from '../movie-card/movie-card.jsx';
-import {GenreMap, VIDEO_PLAY_DELAY_TIME} from '../../constants/constants';
+import {GenreMap} from '../../constants/constants';
 import {connect} from 'react-redux';
 import {ActionCreator} from '../../action';
 import {filterMovies} from '../../helpers/filterMovies';
@@ -10,11 +10,6 @@ export class MoviesList extends PureComponent {
   constructor(props) {
     super(props);
     props.setMovies();
-    this.state = {
-      activeMovieCard: null
-    };
-    this._mouseEnterHandler = this._mouseEnterHandler.bind(this);
-    this._mouseLeaveHandler = this._mouseLeaveHandler.bind(this);
   }
 
   render() {
@@ -28,28 +23,20 @@ export class MoviesList extends PureComponent {
   }
 
   _getMovie(movie) {
-    const {onClick} = this.props;
-    const isPlaying = movie.id === this.state.activeMovieCard;
+    const {onClick, activeItem, onMouseEnter, onMouseLeave} = this.props;
+    const isPlaying = movie.id === activeItem;
+    const _onMouseEnter = () => onMouseEnter(movie.id);
 
     return (
       <MovieCard
         key={movie.id}
         isPlaying={isPlaying}
-        onMouseLeave={this._mouseLeaveHandler}
-        onMouseEnter={this._mouseEnterHandler}
+        onMouseLeave={onMouseLeave}
+        onMouseEnter={_onMouseEnter}
         onClick={onClick}
         {...movie}
       />
     );
-  }
-
-  _mouseEnterHandler(activeMovieCard) {
-    this._timerId = setTimeout(() => this.setState({activeMovieCard}), VIDEO_PLAY_DELAY_TIME);
-  }
-
-  _mouseLeaveHandler() {
-    clearTimeout(this._timerId);
-    this.setState({activeMovieCard: null});
   }
 }
 
@@ -60,7 +47,14 @@ MoviesList.propTypes = {
     poster: PropTypes.string.isRequired
   })),
   setMovies: PropTypes.func.isRequired,
-  onClick: PropTypes.func.isRequired
+  onClick: PropTypes.func.isRequired,
+  onMouseLeave: PropTypes.func.isRequired,
+  onMouseEnter: PropTypes.func.isRequired,
+  activeItem: PropTypes.oneOfType([PropTypes.number, PropTypes.object])
+};
+
+MoviesList.defaultProps = {
+  activeItem: null
 };
 
 const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
