+++ src/components/video-player/video-player.jsx
@@ -1,83 +1,32 @@
-import React, {PureComponent, createRef} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {DefaultVideoSize} from '../../constants/constants';
 
-export class VideoPlayer extends PureComponent {
-  constructor(props) {
-    super(props);
-    this._videoRef = createRef();
-    this.state = {
-      isPlaying: props.isPlaying
-    };
-  }
+export const VideoPlayer = (props) => {
+  const {
+    poster,
+    src,
+    muted,
+    autoPlay,
+    width,
+    height,
+    controls,
+    reference
+  } = props;
 
-  componentDidMount() {
-    const {src} = this.props;
-    const video = this._videoRef.current;
-    this._initCurrentPlayer({video, src});
-  }
-
-  componentDidUpdate() {
-    const video = this._videoRef.current;
-
-    if (this.props.isPlaying) {
-      video.play();
-    } else {
-      video.pause();
-      video.load();
-    }
-  }
-
-  componentWillUnmount() {
-    const video = this._videoRef.current;
-    this._resetCurrentPlayer(video);
-  }
-
-  render() {
-    const {
-      poster,
-      src,
-      muted,
-      autoPlay,
-      width,
-      height,
-      controls
-    } = this.props;
-
-    return (
-      <video
-        poster={poster}
-        width={`${width}px`}
-        height={`${height}px`}
-        controls={controls}
-        src={src}
-        muted={muted}
-        autoPlay={autoPlay}
-        ref={this._videoRef}
-      />
-    );
-  }
-
-  _initCurrentPlayer({video, src}) {
-    video.src = src;
-
-    video.onplay = () => {
-      this.setState({
-        isPlaying: true,
-      });
-    };
-
-    video.onpause = () => this.setState({
-      isPlaying: false,
-    });
-  }
-
-  _resetCurrentPlayer(video) {
-    video.onplay = null;
-    video.onpause = null;
-    video.src = ``;
-  }
-}
+  return (
+    <video
+      poster={poster}
+      width={`${width}px`}
+      height={`${height}px`}
+      controls={controls}
+      src={src}
+      muted={muted}
+      autoPlay={autoPlay}
+      ref={reference}
+    />
+  );
+};
 
 VideoPlayer.propTypes = {
   poster: PropTypes.string.isRequired,
@@ -87,7 +36,8 @@ VideoPlayer.propTypes = {
   width: PropTypes.number,
   height: PropTypes.number,
   isPlaying: PropTypes.bool,
-  controls: PropTypes.bool
+  controls: PropTypes.bool,
+  reference: PropTypes.object.isRequired
 };
 
 VideoPlayer.defaultProps = {
